+++ src/MBrace.Azure/AssemblyInfo.fs
@@ -6,12 +6,12 @@ open System.Reflection
 [<assembly: AssemblyCompanyAttribute("Nessos Information Technologies")>]
 [<assembly: AssemblyCopyrightAttribute("© Nessos Information Technologies.")>]
 [<assembly: AssemblyTrademarkAttribute("MBrace")>]
-[<assembly: AssemblyMetadataAttribute("Release Signature","Version 1.0.0, Git Hash b7f787, Build Date 16112015 11:15 +02:00")>]
-[<assembly: AssemblyVersionAttribute("1.0.0")>]
-[<assembly: AssemblyFileVersionAttribute("1.0.0")>]
+[<assembly: AssemblyMetadataAttribute("Release Signature","Version 1.1.5, Git Hash 7fc047, Build Date 23112015 15:30 +02:00")>]
+[<assembly: AssemblyVersionAttribute("1.1.5")>]
+[<assembly: AssemblyFileVersionAttribute("1.1.5")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.0.0"
+    let [<Literal>] Version = "1.1.5"
 
-    let [<Literal>] ReleaseTag = "1.0.0"
+    let [<Literal>] ReleaseTag = "1.1.5"